Improve responsive design
In this commit, media query breakpoints and width changes were made to collapse and scale the ProductGrid and ProductFiltering components. The ProductFiltering component was also updated to collapse into a dropdown menu for smaller devices, and the mobile navigation menu was changed to a flex column layout.

Furthermore, the ProductFiltering component was enhanced to take into account the discounted price when filtering by price, and several color and font spacing UI improvements were implemented.