Refactor ProductGrid component The ProductGrid component and some of its childs were refactor to improve the code's maintainability. The orderProducts() function was renamed to validate() and was also updated to accommodate changes in the ProductFiltering component.
The ProductFiltering component no longer returns a object with all the products if no filtered selection has been made. This improves the code’s predictability as it matches the same behavior of the ProductSorting component.

The ProductSorting component was also updated to fix a sorting issue after implementing the REST API.